+++ activerecord/lib/active_record/connection_adapters/abstract/connection_handler.rb
@@ -88,14 +88,6 @@ module ActiveRecord
         connection_name_to_pool_manager.keys
       end
 
-      def all_connection_pools
-        ActiveRecord.deprecator.warn(<<-MSG.squish)
-          The `all_connection_pools` method is deprecated in favor of `connection_pool_list`.
-          Call `connection_pool_list(:all)` to get the same behavior as `all_connection_pools`.
-        MSG
-        connection_name_to_pool_manager.values.flat_map { |m| m.pool_configs.map(&:pool) }
-      end
-
       # Returns the pools for a connection handler and  given role. If +:all+ is passed,
       # all pools belonging to the connection handler will be returned.
       def connection_pool_list(role = nil)